requirements.txt
Flask==3.0.3
pdfplumber==0.11.0
PyPDF2==3.0.1
Werkzeug==3.0.3
fpdf2==2.7.9
